Check if current user exists in the database, if not, create new, add exception rescue'ing for loading YAML files.

+++ api_browser.rb
@@ -12,15 +12,22 @@ Dir.glob('app/models/*.rb').each do |model|
 end
 
 begin
-  GitHub::User.delete_all
-  $user = GitHub::User.create(YAML::load_file('config/secrets.yml')[:user])
+  loaded_user = GitHub::User.new(YAML::load_file('config/secrets.yml')[:user])
+  if GitHub::User.last
+    if GitHub::User.last.login == loaded_user.login && GitHub::User.last.token == loaded_user.token
+      $user = GitHub::User.last
+    end
+  else
+    GitHub::User.delete_all
+    $user = GitHub::User.create(YAML::load_file('config/secrets.yml')[:user])
+  end
   
-rescue ActiveRecord::StatementInvalid
+rescue ActiveRecord::StatementInvalid, Errno::ENOENT
+  #TODO: Switching and doing proper actions for single error
+  puts "Could not load config/secrets.yml, did you defined it?"
   puts 'Error: Did you run rake db:migrate?'
   puts 'Trying to do it for you...'
   system 'rake db:migrate'
-  retry
-  puts "Retrying after migration"
 end
 
 uri = URI.parse("http://github.com/api/v2/yaml/user/show/#{$user.login}")
